Fixed test-side issues that kept tests from passing

Related work items: #221196

+++ src/ecomm-api-svc/cypress/integration/mutations/returnReason/updateReturnReason.spec.ts
@@ -11,7 +11,7 @@ describe('Mutation: updateReturnReason', () => {
     var extraIds = [] as SupplementalItemRecord[];
     const mutationName = 'updateReturnReason';
     const createName = 'createReturnReason';
-	const deleteMutName = "deleteReturnReason";
+    const deleteMutName = "deleteReturnReason";
     const queryName = "returnReasons";
     const itemPath = 'returnReason';
     const standardMutationBody = `
@@ -22,25 +22,31 @@ describe('Mutation: updateReturnReason', () => {
         }
     `;
 
-	var deleteItemsAfter = undefined as boolean | undefined;
+    var deleteItemsAfter = undefined as boolean | undefined;
+    var alreadyDeleted = false;
+
     before(() => {
-		deleteItemsAfter = Cypress.env("deleteItemsAfter");
-		cy.deleteCypressItems(queryName, deleteMutName);
+        deleteItemsAfter = Cypress.env("deleteItemsAfter");
+        cy.deleteCypressItems(queryName, deleteMutName);
     });
 
-	beforeEach(() => {
+    beforeEach(() => {
         const name = `Cypress ${mutationName} Test #${itemCount}`;
         const input = `{name: "${name}"}`;
         cy.createAndGetId(createName, itemPath, input).then((returnedId: string) => {
             id = returnedId;
             itemCount++;
         });
-	});
+    });
 
     afterEach(() => {
-		if (!deleteItemsAfter) {
-			return;
-		}
+        if (!deleteItemsAfter) {
+            return;
+        }
+        if (alreadyDeleted) {
+            alreadyDeleted = false;
+            return;
+        }
         if (id !== "") {
             // Delete any supplemental items we created
             cy.deleteSupplementalItems(extraIds).then(() => {
@@ -79,8 +85,9 @@ describe('Mutation: updateReturnReason', () => {
                     ${standardMutationBody}
                 }
             }`;
-            cy.mutationDeletedId(id, mutationName, deleteMutName, mutation, itemPath )
-            
+            cy.mutationDeletedId(id, mutationName, deleteMutName, mutation, itemPath).then(() => {
+                alreadyDeleted = true;
+            });
         });
 
         it("Mutation will fail with invalid 'Name' input", () => {
@@ -122,7 +129,7 @@ describe('Mutation: updateReturnReason', () => {
         it("Mutation with all required input and 'customData' input updates item with customData", () => {
             updateCount++;
             const newName = `Cypress ${mutationName} Update ${updateCount}`;
-            const customData = {data: `${itemPath} customData`, canDelete: true};
+            const customData = { data: `${itemPath} customData`, canDelete: true };
             const mutation = `mutation {
                 ${mutationName}(
                     input: {
@@ -158,14 +165,14 @@ describe('Mutation: updateReturnReason', () => {
 
         it("Mutation with all required input and 'customData' input will overwrite the customData on an existing object", () => {
             const name = `Cypress ${mutationName} customData extra`;
-            const customData = {data: `${itemPath} customData`, extraData: ['C', 'Y', 'P', 'R', 'E', 'S', 'S']};
+            const customData = { data: `${itemPath} customData`, extraData: ['C', 'Y', 'P', 'R', 'E', 'S', 'S'] };
             const input = `{name: "${name}", customData: ${toFormattedString(customData)}}`;
             cy.createAndGetId(createName, itemPath, input, "customData").then((createdItem) => {
                 assert.exists(createdItem.id);
                 assert.exists(createdItem.customData);
-                extraIds.push({itemId: createdItem.id, deleteName: deleteMutName, itemName: name, queryName: queryName});
+                extraIds.push({ itemId: createdItem.id, deleteName: deleteMutName, itemName: name, queryName: queryName });
                 const newName = `Cypress ${mutationName} CD extra updated`;
-                const newCustomData = {data: `${itemPath} customData`, newDataField: { canDelete: true }};
+                const newCustomData = { data: `${itemPath} customData`, newDataField: { canDelete: true } };
                 const mutation = `mutation {
                     ${mutationName}(
                         input: {
@@ -199,7 +206,7 @@ describe('Mutation: updateReturnReason', () => {
                 });
             });
         });
-        
+
         it("Mutation will correctly use all input", () => {
             updateCount++;
             const newName = `Cypress ${mutationName} Update ${updateCount}`;
